server/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.13"
Django = "^4.2.5"
graphene = "^3.3"
graphene-django = "^3.2.0"
django-graphql-jwt = "^0.3.4"
serious-django-graphene = "^0.5.1"
django-cors-headers = "^4.2.0"
mysqlclient = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
